중간저장(기능 기준으로 폴더 분리)
