+++ KSchool/Neis_API/__init__.py
@@ -1,0 +1,11 @@
+from KSchool.Neis_API.region import *
+from KSchool.Neis_API.service import *
+
+
+# service
+info = service.info
+meal = service.meal
+schedule = service.schedule
+
+
+del KSchool
