+++ KSchool/Neis_API/service/__init__.py
@@ -1,0 +1,3 @@
+import KSchool.Neis_API.service.meal
+import KSchool.Neis_API.service.info
+import KSchool.Neis_API.service.schedule
