+++ KSchool/Neis_API/service/exceptions.py
