+++ KSchool/Neis_API/service/request/__init__.py
@@ -1,0 +1,3 @@
+import KSchool.Neis_API.service.request.get
+
+get_request = get.get_request
