+++ KSchool/Neis_API/service/request/errors.py
