+++ KSchool/Neis_API/service/schedule.py
@@ -1,5 +1,4 @@
 import requests
-from Neis_API.service.request import get_request
 
 URL = "https://open.neis.go.kr/hub/SchoolSchedule"
