+++ KSchool/website_crawl/__init__.py
@@ -1,0 +1,4 @@
+import KSchool.website_crawl.letters_to_parents
+import KSchool.website_crawl.get
+
+del KSchool
