--- Neis_API/service/__init__.py
@@ -1,3 +1,0 @@
-import Neis_API.service.meal
-import Neis_API.service.info
-import Neis_API.service.schedule
