--- Neis_API/service/request/__init__.py
@@ -1,4 +1,0 @@
-import Neis_API.service.request.get
-
-get_request = get.get_request
-crawl_website = get.crawl_website
