+++ example/meal_test.py
@@ -1,4 +1,4 @@
-import Neis_API
+import KSchool
 
 meals = Neis_API.meal.meal_date(region=Neis_API.R_SEOUL, school_code="7010083", date="20220610")
 for meal in meals:
